[REFACTOR] #6 - Remove RecordViewModel, seperate to MainViewModel & RecordListViewModel
